requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic-settings==2.5.2
alembic==1.13.3
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0

